ACIProxy/__app__/requirements.txt
adal==1.2.5
attrs==20.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
azure-common==1.1.26
azure-core==1.9.0
azure-functions==1.5.0
azure-identity==1.5.0
azure-mgmt-containerinstance==7.0.0
azure-mgmt-core==1.2.2
azure-mgmt-resource==15.0.0
certifi==2020.11.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cffi==1.14.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
chardet==3.0.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cryptography==3.2.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
importlib-metadata==3.1.0; python_version >= "3.6" and python_version < "3.8"
isodate==0.6.0
jsonschema==3.2.0
msal-extensions==0.3.0
msal==1.6.0
msrest==0.6.19
msrestazure==0.6.4
oauthlib==3.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
portalocker==1.7.1
pycparser==2.20; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyjwt==1.7.1
pyrsistent==0.17.3; python_version >= "3.5"
python-dateutil==2.8.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
pywin32==300; platform_system == "Windows"
pyyaml==5.3.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.25.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
six==1.15.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
urllib3==1.26.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
zipp==3.4.0; python_version >= "3.6" and python_version < "3.8"
